rename text detail file to be more specific
